+++ src/test/java/org/mmarini/railways2/swing/MapPanelTest.java
@@ -29,7 +29,7 @@
 package org.mmarini.railways2.swing;
 
 import org.mmarini.railways2.model.StationStatus;
-import org.mmarini.railways2.model.blocks.BlockStationBuilder;
+import org.mmarini.railways2.model.blocks.BlockBuilder;
 import org.mmarini.railways2.model.blocks.StationDef;
 import org.mmarini.yaml.Utils;
 import org.mmarini.yaml.schema.Locator;
@@ -59,8 +59,9 @@ class MapPanelTest {
                         .build()
                  */
 
-                new BlockStationBuilder(StationDef.create(
-                        Utils.fromResource("/stations/downville.station.yml"), Locator.root()), GAME_DURATION, FREQUENCY, null, null).build()
+                new BlockBuilder(StationDef.create(
+                        Utils.fromResource("/stations/downville.station.yml"), Locator.root()))
+                        .buildStatus(GAME_DURATION, FREQUENCY, null, null)
         ).run();
     }