+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased] yyyy-mm-dd
 
+### Fixed
+
+- Issue #89: Wrong switch toggle
+
 ## [2.0.0] 2023-06-29
 
 ### Added